Consolidate validation logic
