+++ tests/expectations/test_conditions.py
@@ -10,7 +10,6 @@ from great_expectations.expectations.conditions import (
     Condition,
     ConditionParserError,
     InvalidConditionTypeError,
-    InvalidParameterTypeError,
     NullityCondition,
     Operator,
     OrCondition,
@@ -174,32 +173,6 @@ class TestColumn:
         result = col.is_not_null()
 
         assert result == NullityCondition(column=col, is_null=False)
-
-    @pytest.mark.parametrize(
-        "operator_func",
-        [
-            pytest.param(
-                lambda col: col == None,  # noqa: E711  # testing invalid syntax
-                id="eq - Linting error - `is None` is pythonic, "
-                "but that only compares to singleton instance",
-            ),
-            pytest.param(
-                lambda col: col != None,  # noqa: E711  # testing invalid syntax
-                id="ne - Linting error - `is None` is pythonic, "
-                "but that only compares to singleton instance",
-            ),
-            pytest.param(lambda col: col < None, id="lt - Nonsense"),
-            pytest.param(lambda col: col <= None, id="le - Nonsense"),
-            pytest.param(lambda col: col > None, id="gt - Nonsense"),
-            pytest.param(lambda col: col >= None, id="ge - Nonsense"),
-        ],
-    )
-    def test_column_operators_with_none_raises_error(self, operator_func):
-        """Test that Column operators with None parameter raise InvalidParameterTypeError."""
-        col = Column(name="status")
-
-        with pytest.raises(InvalidParameterTypeError):
-            operator_func(col)
 
 
 class TestComparisonCondition:
